Final
so this time it works the same, but you can also click on a rectangle
and it interacts with the boxes and it fills them
if box 1 is empty of has an invalid character, then it fills box 1.
Otherwise  if box 2 is empty of has an invalid character, then it fills
box 2. else if both are valid, then it fills box 1.